requirements.txt
flask
flask-sqlalchemy
sqlalchemy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
PyPDF2
Pillow
